WIP: Testing new system to define light position

+++ source/blender/makesdna/DNA_modifier_types.h
@@ -1972,7 +1972,7 @@ typedef enum eGpencilFlip_Flag {
 typedef struct GpencilLightModifierData {
 	ModifierData modifier;
 	struct Object *object;
-	int loc[3];
+	float loc[3];
 	int flag;                    /* flags */
 	float energy;
 	float ambient;
